[week9/mission]미션2 Modal Slice 활용하여, 모달 기능 추가
