+++ munch/Week9/Week9_Mission/src/components/Navbar.tsx
@@ -1,32 +1,32 @@
 import { useEffect } from "react";
 import { FaShoppingCart } from "react-icons/fa";
 import { useDispatch, useSelector } from "../hooks/useCustomRedux";
-import { calculateTotals } from "../slices/cartSlice";
+import { calculateTotals } from "../features/cart/cartSlice";
 
 const Navbar = () => {
-	const { amount, cartItems } = useSelector((state) => state.cart);
-	const dispatch = useDispatch();
+  const { amount, cartItems } = useSelector((state) => state.cart);
+  const dispatch = useDispatch();
 
-	useEffect(() => {
-		dispatch(calculateTotals());
-	}, [dispatch, cartItems]);
+  useEffect(() => {
+    dispatch(calculateTotals());
+  }, [dispatch, cartItems]);
 
-	return (
-		<div className="flex justify-between items-center p-4 bg-gray-800 text-white">
-			<h1
-				onClick={() => {
-					window.location.href = "/";
-				}}
-				className="text-2xl font-semibold cursor-pointer"
-			>
-				Munch
-			</h1>
-			<div className="flex items-center space-x-2">
-				<FaShoppingCart className="text-2xl" />
-				<span className="text-xl font-medium">{amount}</span>
-			</div>
-		</div>
-	);
+  return (
+    <div className="flex justify-between items-center p-4 bg-gray-800 text-white">
+      <h1
+        onClick={() => {
+          window.location.href = "/";
+        }}
+        className="text-2xl font-semibold cursor-pointer"
+      >
+        Munch
+      </h1>
+      <div className="flex items-center space-x-2">
+        <FaShoppingCart className="text-2xl" />
+        <span className="text-xl font-medium">{amount}</span>
+      </div>
+    </div>
+  );
 };
 
 export default Navbar;
