+++ munch/Week9/Week9_Mission/src/components/PriceBox.tsx
@@ -1,26 +1,22 @@
+import { openModal } from "../features/modal/modalSlice";
 import { useDispatch, useSelector } from "../hooks/useCustomRedux";
-import { clearCart } from "../slices/cartSlice";
 
 const PriceBox = () => {
-	const { total } = useSelector((state) => state.cart);
-	const dispatch = useDispatch();
+  const { total } = useSelector((state) => state.cart);
+  const dispatch = useDispatch();
 
-	const handleInitializeCart = () => {
-		dispatch(clearCart());
-	};
-
-	return (
-		<div className="p-12 flex justify-between">
-			<button
-				type="button"
-				onClick={handleInitializeCart}
-				className="border p-4 rounded-md cursor-pointer"
-			>
-				장바구니 초기화
-			</button>
-			<div>총 가격: {total}원</div>
-		</div>
-	);
+  return (
+    <div className="p-12 flex justify-between">
+      <button
+        type="button"
+        onClick={() => dispatch(openModal())}
+        className="border p-4 rounded-md cursor-pointer"
+      >
+        장바구니 초기화
+      </button>
+      <div>총 가격: {total}원</div>
+    </div>
+  );
 };
 
 export default PriceBox;
