+++ munch/Week9/Week9_Mission/src/features/modal/modalSlice.ts
@@ -1,0 +1,28 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+interface ModalState {
+  isOpen: boolean;
+}
+
+const initialState: ModalState = {
+  isOpen: false,
+};
+
+const modalSlice = createSlice({
+  name: "modal",
+  initialState,
+  reducers: {
+    openModal: (state) => {
+      state.isOpen = true;
+    },
+    closeModal: (state) => {
+      state.isOpen = false;
+    },
+  },
+});
+
+export const { openModal, closeModal } = modalSlice.actions;
+
+const modalReducer = modalSlice.reducer;
+
+export default modalReducer;
